pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend="hatchling.build"

[project]
name = "airflow-priority"
authors = [{name = "the airflow-priority authors", email = "3105306+timkpaine@users.noreply.github.com"}]
description="Priority Tags for Airflow Dags"
readme = "README.md"
license = { text = "Apache-2.0" }
version = "0.1.0"
requires-python = ">=3.9"
keywords = []

classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "apache-airflow>=2,<3",
]

[project.urls]
Repository = "https://github.com/timkpaine/airflow-priority"
Homepage = "https://github.com/timkpaine/airflow-priority"

[project.optional-dependencies]
develop = [
    "build",
    "bump-my-version",
    "check-manifest",
    "hatchling",
    "pytest",
    "pytest-cov",
    "ruff",
    "twine",
    "wheel",
    # Plugins
    "boto3",
    "botocore",
    "datadog-api-client",
    "newrelic-telemetry-sdk",
    "slack-sdk",
]
aws = [
    "boto3",
    "botocore",
]
datadog = [
    "datadog-api-client",
]
newrelic = [
    "newrelic-telemetry-sdk",
]
slack = [
    "slack-sdk",
]

[project.scripts]

[project.entry-points."airflow.plugins"]
airflow_priority_newrelic = "airflow_priority.plugins.newrelic:NewRelicPriorityPlugin"
airflow_priority_datadog = "airflow_priority.plugins.datadog:DatadogPriorityPlugin"
airflow_priority_aws_cloudwatch = "airflow_priority.plugins.aws:AWSCloudWatchPriorityPlugin"
airflow_priority_slack = "airflow_priority.plugins.slack:SlackPriorityPlugin"

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "airflow_priority/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.check-manifest]
ignore = [
    ".copier-answers.yml",
    "Makefile",
    "setup.py",
]

[tool.hatch.build]
artifacts = []

[tool.hatch.build.sources]
src = "/"

[tool.hatch.build.targets.sdist]
include = [
    "/airflow_priority",
    "LICENSE",
    "README.md",
]
exclude = [
    "/.github",
    "/.gitattributes",
    "/.gitignore",
]

[tool.hatch.build.targets.wheel]
include = [
    "/airflow_priority",
]
exclude = [
    "/.github",
    "/.gitattributes",
    "/.gitignore",
    "/pyproject.toml",
]

[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = "airflow_priority/tests"

[tool.ruff]
line-length = 150

[tool.ruff.lint.isort]
combine-as-imports = true
default-section = "third-party"
known-first-party = ["airflow_priority"]
section-order = [
    "future",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
